Retransmit to another home server on proxy fail
This is an quick hack that probably introduces a race condition into the
state machine:

* mark_home_server_zombie() extracted from no_response_to_proxied_request()
* call mark_home_server_zombie() after proxying to a virtual
  server, but before retransmitting to a real home server
* use 1/4 of response window instead of 1/4 of zombie period as a zombie
  tolerance window
* retransmission is done by marking failed server as a zombie,
  repopulating request->home_server, removing it from proxy hash and
  calling proxy_request()

Proper implementation would put this change behind an option (e.g.
replace no_response_fail boolean with a triple-value option), and,
instead of modifying the old request, would drop the old request and
create a new one. For the last part, successfully_proxied_request()
would have to be split into parts that only happen once per client
request and parts that happen for every retransmit.